refactor: extract BlogPosts component
