+++ blog/src/theme/BlogPosts/style.module.scss
@@ -1,0 +1,209 @@
+/* stylelint-disable no-descending-specificity */
+@import "../../css/util";
+
+main.normalPage {
+  width: fit-content;
+  display: grid;
+  grid-template-columns: repeat(3, 430px);
+  margin: 5rem auto 0;
+  align-items: start;
+  font-family: apple-system, system-ui, sans-serif;
+
+  article {
+    margin: 0 1.25rem 2.5rem;
+    border-radius: 1rem;
+    overflow: hidden;
+    border: 2px solid transparent;
+    transition: all 0.3s ease-in-out;
+    // filter: drop-shadow(0 1px 1px rgb(0 0 0 / 10%));
+
+    & > a {
+      width: 100%;
+      height: 100%;
+      overflow: hidden;
+
+      img {
+        object-fit: cover;
+        width: 100%;
+        max-width: unset;
+        // filter: drop-shadow(0 1px 0 rgb(0 0 0 / 4%));
+        transform-origin: center center;
+        transition: all 0.3s ease-in-out;
+      }
+    }
+
+    & .content {
+      padding: 0 1rem 1rem;
+
+      header {
+        & .tags {
+          display: flex;
+          justify-content: flex-start;
+          overflow: hidden;
+          text-overflow: ellipsis;
+          margin: 0.8rem 0;
+
+          & > a {
+            color: var(--ifm-link-color);
+            margin-right: 0.5rem;
+            text-transform: uppercase;
+            font-size: 0.66rem;
+
+            &:hover {
+              opacity: 0.6;
+            }
+          }
+        }
+
+        h2 {
+          font-size: 1.375rem;
+          line-height: 1.2em;
+          color: #222;
+          margin-top: 0.8rem;
+          margin-bottom: 0.5em;
+          transition: all 0.3s ease-in-out;
+
+          &:hover {
+            opacity: 0.6;
+          }
+        }
+
+        p {
+          color: #1d1d1f;
+          max-height: #{1.7 * 3}em;
+          overflow-y: hidden;
+          text-overflow: ellipsis;
+          font-weight: 300;
+          margin-bottom: 2rem;
+        }
+      }
+
+      & .footer {
+        display: flex;
+        align-items: center;
+      }
+
+      & .authors {
+        display: flex;
+        flex-direction: row-reverse;
+
+        & .author,
+        & > span {
+          width: 32px;
+          height: 32px;
+          border-radius: 50%;
+          margin-left: -1rem;
+
+          &:last-child {
+            margin-left: 0;
+          }
+        }
+      }
+
+      & .divider {
+        margin: 0 0.5rem;
+      }
+    }
+
+    &:hover {
+      border-color: var(--ifm-color-primary);
+      transform: translateY(-20px);
+
+      & > a {
+        img {
+          transform: scale3d(1.05, 1.05, 1);
+        }
+      }
+
+      h2 {
+        opacity: 0.6;
+      }
+
+      & .tags:hover + a > h2 {
+        opacity: 1;
+      }
+    }
+  }
+}
+
+@include respond-above(sm) {
+  main.firstPage {
+    grid-template-columns: repeat(2, 645px);
+
+    article {
+      & > a {
+        img {
+          width: 100%;
+        }
+      }
+
+      .content {
+        padding: 0 1.875rem 1.875rem;
+      }
+
+      &:first-of-type {
+        grid-column: 1 / 3;
+        display: flex;
+        filter: none;
+        margin-top: 0;
+        margin-bottom: 5rem;
+
+        & > a {
+          border-radius: 1rem;
+          width: 601px;
+          height: auto;
+          flex-shrink: 0;
+
+          img {
+            height: 100%;
+            width: 100%;
+          }
+        }
+
+        .content {
+          padding: 1rem 2rem;
+
+          header {
+            &::before {
+              content: "LATEST POST";
+              font-size: 0.7rem;
+              font-weight: 800;
+            }
+
+            & .tags > a {
+              font-size: 0.8rem;
+            }
+
+            h2 {
+              font-size: 2.55rem;
+            }
+
+            p {
+              font-size: 1.2rem;
+            }
+          }
+        }
+
+        &:hover {
+          border-color: transparent;
+          transform: none;
+
+          & > a {
+            img {
+              transform: scale3d(1.2, 1.2, 1) rotate3d(0, 0, 1, -2deg);
+            }
+          }
+        }
+      }
+    }
+  }
+}
+
+@include respond-below(sm) {
+  main.normalPage {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    margin: 0;
+  }
+}
