+++ blog/src/theme/BlogTagsPostsPage/style.module.scss
@@ -1,5 +1,3 @@
-@import "../BlogListPage/style.module";
-
 .header {
   text-align: center;
   margin-bottom: 4rem;
